fix(): Fix for leagues
